checking in working code gen for structs

+++ smithy-swift/smithy-swift-codegen-test/expected/Weather/models/GetForecastOutput.swift
@@ -14,5 +14,5 @@
 //
 
 struct GetForecastOutput {
-    public let chanceOfRain: Int?
+    public let chanceOfRain: Float?
 }
